Created PostDTO and Resource Exception classes

+++ src/main/java/com/masai/blogapp/entity/Post.java
@@ -27,7 +27,7 @@ public class Post {
 	
 	@Id
 	@GeneratedValue( strategy= GenerationType.IDENTITY )
-	private Integer postID;
+	private Long postID;
 	
 	@NotNull
 	private String title;
